requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
httpx==0.25.0
asyncpg==0.29.0
langchain-openai==0.1.25
langgraph==0.2.34
langgraph-checkpoint==2.0.2
pydantic==2.10.3
torch==2.2.2
torchvision==0.17.2
Pillow
numpy==1.24.3
git+https://github.com/openai/CLIP.git